Handle edge cases in incremental doc rendering
Specifically, if docCount % 5 != 0, just distribute the extra docs into
the first bucket (i.e. for the first render). This’ll never be greater
than 5 docs, of course, so the unevenness is negligible.

And, if there’s less than 5 docs, handle that too by just outputting
the final matrix 5 times. This is sending a bit of extra data over the
wire, but it lets our client be simpler.