Rewrite + a workable force-directed layout
First of all, this is a major clean up of the existing codebase. I
converted it to ES6 (transpiled with tracer by gulp), which provided an
increasingly-needed division of the client-side code into modules,
along with a generally cleaner syntax. I also removed some unneeded
libraries, cleaned up the template, etc.

In terms of new functionality, the force directed layout now works
decently. This took quite a lot of playing.

Nothing that I tried for creating edges (and setting their distances)
before starting the layout simulation worked. The final layout that we
want is more or less a spiral (bigger words in the center, with smaller
words radiating out), which means that two low-frequency words could be
right next to each other (along the outer edge of the spiral) or
completely across the layout from each other, and I wasn’t sure how to
create edges to make this happen.

So, the approach I ended up using simply leaves the graph disconnected
but then gives each word/node a repulsive force proportional to its
frequency (and to the window’s area, so the visualization gets more
breathing room as the window expands, which lowers the odds of word
overlap and is more aesthetically pleasing).

Making this work also required setting the initial conditions of the
layout in such a way that the frequent words would be in the center.
Further, I wanted the initial position of a word to be 1) deterministic
and 2) independent of its index in the token list, so that the layout
would stay largely intact when a token was removed. To accomplish this
I ended up using polar coordinates. I set the theta for each word as
(2π*(frequency % 360)/360). The (frequency % 360)/360 part is
semi-arbitrary, but it nicely satisfies the two conditions listed
above. Then, I set the r of each node/word to be (node’s frequency/max
frequency)*(length of the line segment drawn at angle ø from the center
to the nearest window edge).

Finally, to keep the nodes from spreading out into space too far, I
added another force like the built in gravity option (though mine is
sensitive to the window’s aspect ratio) and imposed a hard bounding so
no node can go off the edge of the screen. This is coupled with code
that reduces the number of words shown as the window’s area shrinks,
again so that all the words can fit without collision. (The word count
is reduced almost exponentially as the window shrinks, rather than
linearly, because the remaining words at a given size are, by
definition, disproportionately large.

The above algorithm works as well as any I’ve able to make so far, but
its layout is still both less compact than the non-force directed
algorithm’s and, unlike the original layout, is prone to word overlap
at far greater rates than I’d like. When I tried to implement a
collision detection algorithm to prevent the overlaps (after adapting
that algorithm to work with word-shaped items, not points or circles),
it didn’t work, because the collision detection algorithm ended up
moving the biggest words way too far from the center (as there was
nothing holding them there; they were there only by virtue of their
initial positions).

After a bit of googling, I stumbled upon a Microsoft paper
(http://research.microsoft.com/en-us/um/people/weiweicu/images/wordpv.pd
f) that points to a solution: draw edges between nodes (with a Delaunay
triangulation) _after_ doing a rough initial layout, so that then those
edges can hold the nodes in place during the collision detection
process. Brilliant! Naturally, though, I found this paper only after
doing all the above work. I will implement it, but not until next week
when I have more time.