Add a compacting force
This replaces d3’s standard links-attract force. It’s better because it
stops exerting a force once the words start overlapping, which d3’s
implementation isn’t smart enough to do. Unfortunately, if making this
new force really strong makes the graph highly non-planar (i.e. far
from its initial layout), so I’ve turned it down for now. But, in the
future, when we have a force that keeps the graph planar, we’ll be able
to crank it back up to awesome effect.