Treat excluded words as a list, in addition to as a set
We need to store the list of excluded words with an order, so that we
can always show the last excluded word at the top of the excluded words
list. We were able to fake this before, but, now that the list is being
persisted across requests, we actually need to store the order.